Using react buttons for changing vehicle
